Separating animation from exotic.py into animate.py and making it able to be accessed across multiple .py files.
